changed folder name from Heaps to Heap. corrected all folder names.
