+++ Heap/Jesse and Cookies/solution.cpp
