+++ Heap/Jesse and Cookies/solution.py
