+++ Heap/QHEAP1/solution.cpp
